refactor: 기타 도메인 패키지 변경 (#133) menu, seating
