+++ src/main/java/site/gachontable/infra/security/principal/UserAuthDetailsService.java
@@ -5,9 +5,9 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import site.gachontable.domain.shared.Role;
-import site.gachontable.domain.user.domain.User;
-import site.gachontable.domain.user.port.out.UserRepository;
-import site.gachontable.domain.user.exception.UserNotFoundException;
+import site.gachontable.domain.member.domain.User;
+import site.gachontable.domain.member.port.out.UserRepository;
+import site.gachontable.domain.member.exception.UserNotFoundException;
 
 @Service
 @RequiredArgsConstructor
